LayoutTests/media/encrypted-media/clearKey/clearKey-message-cenc-event-mse-expected.txt


EME API is supported OK
Media source is opened OK
EVENT(encrypted)
MediaKeys is created OK
EVENT(message)
Expected Kid OK
END OF TEST
